feat: Update CustomExceptionHandler
